backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
databricks-sql-connector

# Development dependencies
black
isort
flake8
mypy
pytest
pytest_asyncio
httpx
